python_optimizer/pyproject.toml
[project]
name = "axdspy-gepa-optimizer"
version = "0.1.0"
description = "Flask server exposing dspy.GEPA optimizer"
requires-python = ">=3.10"
dependencies = [
  "flask>=3.0.0",
  "dspy-ai>=2.5.0",
]

[tool.uv]
dev-dependencies = []

